fix: :recycle: move Layout in shared/components
